NodeDefCode: added useRequest hook to simplify useItems hook
